09-24 Appt Ownerships

+++ controllers/appointment.js
@@ -58,7 +58,13 @@ exports.getAppointment = async (req, res, next) => {
 // @access   Private
 exports.addAppointment = async (req, res, next) => {
     try {
+        req.body.user = req.user.id;
         req.body.hospital = req.params.hospitalId;
+
+        const exisedAppointment = await Appointment.find({user:req.user.id});
+        if (exisedAppointment.length >= 3 && req.user.role !== 'admin') {
+            return res.status(400).json({ success: false, msg: `The user with id ${userId} has already made 3 appointments.` });
+        }
 
         const hospital = await Hospital.findById(req.params.hospitalId);
 
@@ -86,6 +92,10 @@ exports.updateAppointment = async (req, res, next) => {
             return res.status(404).json({ success: false, msg: `No appointment with the ID of ${id} found.` });
         }
 
+        if (appointment.user.toString() !== req.user.id && req.user.role !== 'admin') {
+            return res.status(401).json({ success: false, msg: `User ${req.user.id} is not authorized to update this appointment` });
+        }
+
         appointment = await Appointment.findByIdAndUpdate(req.params.id, req.body, {
             new: true,
             runValidators: true
@@ -107,6 +117,10 @@ exports.deleteAppointment = async (req, res, next) => {
 
         if (!appointment) {
             return res.status(404).json({ success: false, msg: `No appointment with the ID of ${id} found.` });
+        }
+
+        if (appointment.user.toString() !== req.user.id && req.user.role !== 'admin') {
+            return res.status(401).json({ success: false, msg: `User ${req.user.id} is not authorized to delete this appointment` });
         }
 
         await appointment.deleteOne();
